Release notes: show version number for "Changes since" section
Add the current version number as a prefix to the 'Changes since' line.

The "Changes since" sections of the release notes are 'below the fold'
so it is not immediately obvious to readers what is the version the
changes refer to. The version is allready available as the header id
but this isn't displayed to the user in any obvious manner.

The html release notes processing is multi-phase. There are maintainer
actions to add new notes to the original markdown version. This is
followed by conversion of the markdown to html, and finally some format
adjustments to the html in render-release-notes.sh.

This code change is to the final html adjustments to avoid any changes
to the maintainer workflow.

The updated perl script isn't quite as neat as previously,
but is obvious.

Signed-off-by: Philip Oakley <philipoakley@iee.email>