SDK update-via-packman: add progress notifications
When upgrading the SDK it is suprising that git-extra is always
re-installed, confusing developers not yet familiar with the
convoluted ways needed for the SDK update.

Tell the user of the two process steps that are happening.

This change may need extra changes in the pacman repository as well
to complete the cycle.

Signed-off-by: Philip Oakley <philipoakley@iee.email>